pyproject.toml
[project]
name = "pr-guardian"
version = "0.1.0"
description = "todo: make a description"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-mistralai>=1.1.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
